+++ src/datasources/work-items/components/WorkItemsQueryEditor.test.tsx
@@ -21,6 +21,14 @@ jest.mock('./query-builder/WorkItemsQueryBuilder', () => ({
 }));
 
 describe('WorkItemsQueryEditor', () => {
+  it('should call onRunQuery on init', () => {
+    const render = setupRenderer(WorkItemsQueryEditor, WorkItemsDataSource);
+
+    const [, onRunQuery] = render({});
+
+    expect(onRunQuery).toHaveBeenCalledTimes(1);
+  });
+
   let getCustomPropertyOptionsSpy: jest.SpyInstance;
 
   beforeEach(() => {
@@ -101,6 +109,7 @@ describe('WorkItemsQueryEditor', () => {
       try {
         const render = setupRenderer(WorkItemsQueryEditor, WorkItemsDataSource);
         const [onChange, onRunQuery] = render({ types: [WorkItemTypeOptions.WorkOrders] });
+        onRunQuery.mockClear();
 
         await userEvent.click(page.removeOptionButton('Work orders'));
 
@@ -128,6 +137,7 @@ describe('WorkItemsQueryEditor', () => {
       try {
         const render = setupRenderer(WorkItemsQueryEditor, WorkItemsDataSource);
         const [onChange, onRunQuery] = render({ properties: [WorkItemPropertiesOptions.ID] });
+        onRunQuery.mockClear();
 
         await userEvent.click(page.removeOptionButton('Work item ID'));
 
@@ -153,6 +163,7 @@ describe('WorkItemsQueryEditor', () => {
       const render = setupRenderer(WorkItemsQueryEditor, WorkItemsDataSource);
 
       const [onChange, onRunQuery] = render({});
+      onRunQuery.mockClear();
 
       page.setTakeLimit('-5');
 
@@ -165,6 +176,7 @@ describe('WorkItemsQueryEditor', () => {
       const render = setupRenderer(WorkItemsQueryEditor, WorkItemsDataSource);
 
       const [onChange, onRunQuery] = render({});
+      onRunQuery.mockClear();
 
       page.setTakeLimit(`${TAKE_LIMIT + 1}`);