+++ src/datasources/work-items/components/WorkItemsQueryEditor.tsx
@@ -61,6 +61,7 @@ export function WorkItemsQueryEditor({ query, onChange, onRunQuery, datasource }
   const isTypesValid = isTypesNonEmpty(query.types);
   const takeInvalidMessage = getTakeError(query.take);
   const isTakeValid = takeInvalidMessage === '';
+  const outputType = query.outputType ?? OutputType.Properties;
 
   const [customPropertyOptions, setCustomPropertyOptions] = useState<Array<ComboboxOption<string>>>([]);
   const [isCustomPropertiesInitialized, setIsCustomPropertiesInitialized] = useState(false);
@@ -214,6 +215,13 @@ export function WorkItemsQueryEditor({ query, onChange, onRunQuery, datasource }
     [onChange, onRunQuery]
   );
 
+  useEffect(() => {
+    if (!query.outputType) {
+      handleQueryChange({ ...query, outputType: OutputType.Properties });
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
   const onOutputTypeChange = (value: OutputType) => {
     handleQueryChange({ ...query, outputType: value });
   };
@@ -269,7 +277,7 @@ export function WorkItemsQueryEditor({ query, onChange, onRunQuery, datasource }
         <RadioButtonGroup
           options={outputTypeOptions}
           onChange={onOutputTypeChange}
-          value={query.outputType}
+          value={outputType}
         />
       </InlineField>
       <InlineField
@@ -290,7 +298,7 @@ export function WorkItemsQueryEditor({ query, onChange, onRunQuery, datasource }
           maxWidth={CONTROL_WIDTH}
         />
       </InlineField>
-      {query.outputType === OutputType.Properties && (
+      {outputType === OutputType.Properties && (
         <>
           <InlineField
             label={labels.properties}
@@ -329,7 +337,7 @@ export function WorkItemsQueryEditor({ query, onChange, onRunQuery, datasource }
             onChange={onFilterChange}
           />
         </InlineField>
-        {query.outputType === OutputType.Properties && (
+        {outputType === OutputType.Properties && (
            <Stack direction="column" gap={0}>
               <InlineField
                 label={labels.orderBy}